% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Build Ground-Motion Data Products from OpenQuake}
\usage{
buildGMDP(
  path,
  IDo = "00000000",
  engine = "openquake",
  vs30 = NULL,
  vref,
  quantile_AF = "mean",
  param = FALSE
)
}
\arguments{
\item{path}{character. Path to the folder containing OQ hazard & UHS CSVs/zips}

\item{IDo}{character Output ID label}

\item{engine}{character c("openquake","user")}

\item{vs30}{numeric. Vs30 steps}

\item{vref}{numeric reference Vs30}

\item{quantile_AF}{character. c("mean", 0.16, 0.50, 0.84)}

\item{param}{logical. If TRUE => do param expansions}
}
\value{
A list of \code{(AEPTable, UHSTable, AFmodel_AEP, AFmodel_UHS, SaTRmodel, RMwTable)}
}
\description{
Replicates the old code logic but allows \code{param=FALSE} to skip param expansions.
Always reads "as-is" UHS from CSV if present, so \code{UHSTable} is never empty.
}
