% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.R
\name{buildHCPlot}
\alias{buildHCPlot}
\title{Title}
\usage{
buildHCPlot(
  HEIGHT = 600,
  LEGEND = TRUE,
  TIP = "ID:{point.series.name}<br> X={point.x}<br> Y={point.y}",
  CURVE = NULL,
  BAR = NULL,
  POINT = NULL,
  COLUMN = NULL,
  XT = "X []",
  YT = "Y []",
  COLORS = hcl.colors(n = 5, palette = "ag_Sunset", alpha = 0.6),
  TYPE = "line",
  XLOG = TRUE,
  YLOG = FALSE,
  XREV = FALSE,
  YREV = FALSE,
  XMAX = NULL,
  YMAX = NULL,
  XMIN = NULL,
  YMIN = NULL,
  LAYOUT = "horizontal",
  ALIGN = "right",
  VALIGN = "top",
  LINE = "Solid",
  THEME = hc_theme_hcrt(),
  SYMBOL = "circle"
)
}
\arguments{
\item{HEIGHT}{Integer. Height of the plot in pixels}

\item{LEGEND}{Boolean. Show legend}

\item{TIP}{String. Tooltip}

\item{CURVE}{data.table. Data for the curve}

\item{BAR}{data.table. Data for the bar}

\item{POINT}{data.table. Data for the point}

\item{COLUMN}{data.table. Data for the column}

\item{XT}{String. X axis title}

\item{YT}{String. Y axis title}

\item{COLORS}{Vector. Colors for the plot}

\item{TYPE}{String. Type of the plot}

\item{XLOG}{Boolean. Logarithmic scale for X axis}

\item{YLOG}{Boolean. Logarithmic scale for Y axis}

\item{XREV}{Boolean. Reverse X axis}

\item{YREV}{Boolean. Reverse Y axis}

\item{XMAX}{Double. Maximum value for X axis}

\item{YMAX}{Double. Maximum value for Y axis}

\item{XMIN}{Double. Minimum value for X axis}

\item{YMIN}{Double. Minimum value for Y axis}

\item{LAYOUT}{String. Layout of the legend}

\item{ALIGN}{String. Alignment of the legend}

\item{VALIGN}{String. Vertical alignment of the legend}

\item{LINE}{String. Line style}

\item{THEME}{String. Theme}

\item{SYMBOL}{String. Symbol}
}
\value{
Highchart. Highchart object
}
\description{
Title
}
