% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.ASCE.R
\name{buildTable.ASCE}
\alias{buildTable.ASCE}
\title{Title}
\usage{
buildTable.ASCE(units = "m", engine = "flextable", size = 12)
}
\arguments{
\item{units}{c("m","m/s","ft","ft/s")}

\item{engine}{c("flextable","kableExtra","gtable")}

\item{size}{Font Size}
}
\value{
data.table
}
\description{
Title
}
\examples{
buildTable.ASCE()
buildTable.ASCE(units="m/s")
buildTable.ASCE(units="m/s",engine="flextable")
buildTable.ASCE(units="m/s",size=8)
}
