% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.Kmax.R
\name{buildTable.Kmax}
\alias{buildTable.Kmax}
\title{Build Seismic Coefficient Table (Kmax) in g}
\usage{
buildTable.Kmax(
  .x,
  Tso,
  Dao,
  Vs30o,
  size = 12,
  po = "mean",
  TRo = c(500, 1000, 2500, 5000, 10000),
  tagUnits = FALSE
)
}
\arguments{
\item{.x}{UHSTable object}

\item{Tso}{numeric Ts in seconds}

\item{Dao}{numeric. Da in cm}

\item{Vs30o}{number}

\item{size}{font size}

\item{po}{quantiles}

\item{TRo}{return periods in years}

\item{tagUnits}{boolean}
}
\value{
Table
}
\description{
Build Seismic Coefficient Table (Kmax) in g
}
