% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.Sa.R
\name{buildTable.Sa}
\alias{buildTable.Sa}
\title{Title}
\usage{
buildTable.Sa(
  x,
  Tno = NULL,
  size = 12,
  po = c(0.16, 0.5, 0.84),
  engine = "flextable",
  TRo = c(500, 1000, 2500, 5000, 10000),
  Vs30o = NULL,
  SIDo = NULL
)
}
\arguments{
\item{x}{GMDP object}

\item{Tno}{Period in seconds}

\item{size}{Font size}

\item{po}{c("mean",0.16,0.50,0.84)}

\item{engine}{c("flextable","kableExtra")}

\item{TRo}{Return Period in years}

\item{Vs30o}{Vs30 in m/s}

\item{SIDo}{Site Class}
}
\value{
Table
}
\description{
Title
}
