% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.Ts.R
\name{buildTable.Ts}
\alias{buildTable.Ts}
\title{Slope's Fundamental Period Table}
\usage{
buildTable.Ts(.geometry, .material, regression = "qrf")
}
\arguments{
\item{.geometry}{Data frame}

\item{.material}{Data frame}

\item{regression}{String. Regression method}
}
\value{
List
}
\description{
Slope's Fundamental Period Table
}
