% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFractions.R
\name{checkFractions}
\alias{checkFractions}
\title{Title}
\usage{
checkFractions(GravelsFraction, SandsFraction, FinesFraction)
}
\arguments{
\item{GravelsFraction}{Integer. Gravels fraction (0-100) in \%}

\item{SandsFraction}{Integer. Sands fraction (0-100) in \%}

\item{FinesFraction}{Integer. Fines fraction (0-100) in \%
#'}
}
\value{
Boolean. TRUE if fractions are OK, FALSE otherwise
}
\description{
Title
}
