% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CylinderRoots.R
\name{CylinderRoots}
\alias{CylinderRoots}
\title{Cylinder Roots Data}
\format{
A data frame with 307124 rows and 4 variables:
\describe{
\item{n}{Integer: Root number (1-10)}
\item{m}{Double: Inhomogeneity ratio (0-0.95)}
\item{l}{Double: Aspect Ratio (0-0.4950)}
\item{an}{Double: Cylinder roots}
}
}
\usage{
data(CylinderRoots)
}
\description{
This dataset contains cylinder roots.
}
