% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parseUHScolumn}
\alias{.parseUHScolumn}
\title{Parse UHS Column Name (Internal)}
\usage{
.parseUHScolumn(tag)
}
\arguments{
\item{tag}{Character, e.g. \code{"0.095160~SA(0.2)"}.}
}
\value{
A two-element list with \verb{(rate, Tn)}. Tn=0 if it detects "PGA."
}
\description{
A helper that splits a column name like \code{0.095160~PGA} or \code{0.048770~SA(0.2)}
into numeric rate and Tn.
}
\keyword{internal}
