% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.AF.TR.R
\name{fitModel.AF.TR}
\alias{fitModel.AF.TR}
\title{Fit Amplification Factor Model for Return Period}
\usage{
fitModel.AF.TR(.x, pga, Tn, vs30, vref, q = 0.5, Vl = 200, Vu = 2000)
}
\arguments{
\item{.x}{Data.table. Input data containing spectral acceleration values.}

\item{pga}{Numeric. Peak ground acceleration value.}

\item{Tn}{Numeric. Natural period.}

\item{vs30}{Numeric. Target Vs30 value.}

\item{vref}{Numeric. Reference Vs30 value (must be either 760 or 3000 m/s).}

\item{q}{Character or numeric. Quantile for amplification factors. Default is 0.50.}

\item{Vl}{Numeric. Lower bound for Vs30. Default is 200 m/s.}

\item{Vu}{Numeric. Upper bound for Vs30. Default is 2000 m/s.}
}
\value{
A data.table containing:
\itemize{
\item PGA: Peak ground acceleration
\item Vref: Reference Vs30 value
\item Vs30: Target Vs30 value
\item AF: Amplification factor
\item sdLnAF: Standard deviation of the natural logarithm of the amplification factor
}
}
\description{
Fits an amplification factor model for a given return period based on Vs30 and reference Vs30 values.
}
