% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.AFTR.R
\name{fitModel.AFTR}
\alias{fitModel.AFTR}
\title{Amplification Factor AF, from Stewart Et Al. 2016}
\usage{
fitModel.AFTR(x, Vs30 = NULL, Vref = NULL, Vl = 200, Vu = 2000)
}
\arguments{
\item{x}{GMDP object}

\item{Vs30}{Vs30 in m/s}

\item{Vref}{Reference Vs30 in m/s}

\item{Vl}{Lower limit Vs30 in m/s}

\item{Vu}{Upper limit Vs30 in m/s}
}
\value{
data.table
}
\description{
Amplification Factor AF, from Stewart Et Al. 2016
}
