% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.an.R
\name{fitModel.an}
\alias{fitModel.an}
\title{Title}
\usage{
fitModel.an(no = 1, mo, lo, OSF = 0.3)
}
\arguments{
\item{no}{Integer. Number of mode n=1..8}

\item{mo}{Double. Inhomogeinity ratio. mo=0: homogeneous material. 0<=mo<=0.95}

\item{lo}{Double. Truncation Ratio lo<=0.5. no=0: no truncation. (bo=0)}

\item{OSF}{Double. Overshooting Factor. Default value: 0.30 (30\%)}
}
\value{
Number. Cylinder roots
}
\description{
Title
}
