% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Dn.TR.R
\name{fitModel.Dn.TR}
\alias{fitModel.Dn.TR}
\title{Title}
\usage{
fitModel.Dn.TR(
  Tn,
  Sa,
  PGA,
  xD = 1.3,
  Mw = 6.5,
  kymin = 0.005,
  kymax = 0.5,
  n = 30
)
}
\arguments{
\item{Tn}{Double. Natural Period}

\item{Sa}{Double. Spectral Acceleration}

\item{PGA}{Double. Peak Ground Acceleration}

\item{xD}{Double. BM19 Factor}

\item{Mw}{Double. Magnitude}

\item{kymin}{Double. Minimum ky value}

\item{kymax}{Double. Maximum ky value}

\item{n}{Integer. Number of ky values}
}
\value{
data.table
}
\description{
Title
}
