% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.DnTR.R
\name{fitModel.DnTR}
\alias{fitModel.DnTR}
\title{Title}
\usage{
fitModel.DnTR(x, xD = 1.3, Mw = 6.5, kymin = 0.005, kymax = 0.5, n = 30)
}
\arguments{
\item{x}{GMDP object}

\item{xD}{BM19 Factor}

\item{Mw}{Magnitude}

\item{kymin}{Minimum ky value}

\item{kymax}{Maximum ky value}

\item{n}{Number of ky values}
}
\value{
data.table
}
\description{
Title
}
