% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Kmax.TR.R
\name{fitModel.Kmax.TR}
\alias{fitModel.Kmax.TR}
\title{Title}
\usage{
fitModel.Kmax.TR(a, b, e, PGA, Ts, n = 20)
}
\arguments{
\item{a}{numeric. a parameter}

\item{b}{numeric. b parameter}

\item{e}{numeric. e parameter}

\item{PGA}{numeric. PGA value}

\item{Ts}{numeric. Ts value}

\item{n}{numeric. Number of points}
}
\value{
data.table
}
\description{
Title
}
