% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.KmaxTR.R
\name{fitModel.KmaxTR}
\alias{fitModel.KmaxTR}
\title{Title}
\usage{
fitModel.KmaxTR(x, n = 20)
}
\arguments{
\item{x}{data.table}

\item{n}{Number of points}
}
\value{
data.table
}
\description{
Title
}
