% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Title}
\usage{
fitModel(
  .data,
  x = NULL,
  y,
  .newdata = NULL,
  level = 0.95,
  regression = "qrf",
  removeZeroInstances = FALSE,
  uniqueResponses = FALSE,
  ntree = 500
)
}
\arguments{
\item{.data}{data.table}

\item{x}{character}

\item{y}{character}

\item{.newdata}{vector}

\item{level}{double.}

\item{regression}{character}

\item{removeZeroInstances}{logical}

\item{uniqueResponses}{logical}

\item{ntree}{integer}
}
\value{
vector
}
\description{
Title
}
