% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Shear.R
\name{fitModel.Shear}
\alias{fitModel.Shear}
\title{Title}
\usage{
fitModel.Shear(.data = SiteTable, .newdata, level = "mean", regression = "qrf")
}
\arguments{
\item{.data}{Double. Data frame}

\item{.newdata}{Data frame. New data to predict}

\item{level}{Double. Confidence level (0-1)}

\item{regression}{String. Regression method}
}
\value{
List
}
\description{
Title
}
