% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Td.R
\name{fitModel.Td}
\alias{fitModel.Td}
\title{Title}
\usage{
fitModel.Td(.geometry, .material, level = "mean", regression = "qrf")
}
\arguments{
\item{.geometry}{Data frame}

\item{.material}{Data frame}

\item{level}{Double. Confidence level (0-1)}

\item{regression}{String. Regression method}
}
\value{
List
}
\description{
Title
}
