% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.TdF.R
\name{fitModel.TdF}
\alias{fitModel.TdF}
\title{Title}
\usage{
fitModel.TdF(
  Hmax,
  lo = NULL,
  Bmax = NULL,
  Bo = NULL,
  GravelsFraction = NULL,
  SandsFraction = NULL,
  FinesFraction = NULL,
  level = 0.95
)
}
\arguments{
\item{Hmax}{Double. Height in m}

\item{lo}{Double. Truncation ratio}

\item{Bmax}{Double. Maximum width in m}

\item{Bo}{Double. TSF Crest in m}

\item{GravelsFraction}{Integer. Gravels fraction (0-100) in \%}

\item{SandsFraction}{Integer. Sands fraction (0-100) in \%}

\item{FinesFraction}{Integer. Fines fraction (0-100) in \%}

\item{level}{Double. Confidence level (0-1)}
}
\value{
List
}
\description{
Title
}
