% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Ts.R
\name{fitModel.Ts}
\alias{fitModel.Ts}
\title{Title}
\usage{
fitModel.Ts(VSm, hs, zm)
}
\arguments{
\item{VSm}{Vector. Shear Wave Velocity profile in m/s (vector)}

\item{hs}{Vector. Layer Thickness profile in m (vector)}

\item{zm}{Vector. Layer Coordinates profile in m (vector)}
}
\value{
Double. Site's fundamental period Ts in seconds
}
\description{
Title
}
