% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.VSo.F.R
\name{fitModel.VSo.F}
\alias{fitModel.VSo.F}
\title{Title}
\usage{
fitModel.VSo.F(
  Hso,
  GravelsFraction = NULL,
  SandsFraction = NULL,
  FinesFraction = NULL,
  OSF = 0.3,
  POPo = 0,
  WaterFraction = 0,
  level = 0.95
)
}
\arguments{
\item{Hso}{Double. Height of the soil column (m)}

\item{GravelsFraction}{Integer. Gravels fraction (0-100) in \%}

\item{SandsFraction}{Integer. Sands fraction (0-100) in \%}

\item{FinesFraction}{Integer. Fines fraction (0-100) in \%}

\item{OSF}{Double. Overshooing Factor}

\item{POPo}{Double. Pre-Overburden Pressure in kPa}

\item{WaterFraction}{Dowuble. Water Table in \% as Hw=Water*Hs}

\item{level}{Double. Confidence level (0-1)}
}
\value{
Double. Maximum Shear Module (MPA)
}
\description{
Title
}
