% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Td.R
\name{get.Td}
\alias{get.Td}
\title{Title}
\usage{
get.Td(.geometry, .material, regression = "qrf")
}
\arguments{
\item{.geometry}{Data frame}

\item{.material}{Data frame}

\item{regression}{String. Regression method}

\item{level}{Double. Confidence level (0-1)}
}
\value{
List
}
\description{
Title
}
