% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmdp-package.R
\docType{package}
\name{gmdp-package}
\alias{gmdp}
\alias{gmdp-package}
\title{gmdp: Ground-Motion Design Parameters}
\description{
The GMDP library provides tools for estimating design parameters in seismic geotechnical engineering. The package includes tools for importing seismic hazard data from various platforms (openQuake, FRisk), models for estimating the dynamic response of sites using spectral amplification models, models for estimating intensity measures in terms of ground accelerations and pseudo-accelerations on firm soil, generation of soil profiles through simulation, and models for estimating the site's dynamic properties.
}
\author{
\strong{Maintainer}: Alejandro Verri Kozlowski \email{averri@fi.uba.ar} (\href{https://orcid.org/0000-0002-8535-1170}{ORCID})

}
\keyword{internal}
