% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.userAEP.R
\name{importModel.userAEP}
\alias{importModel.userAEP}
\title{Import AEP values from user table in Excel}
\usage{
importModel.userAEP(path = NULL, filename = "AEP.xlsx")
}
\arguments{
\item{path}{Path to the file}

\item{filename}{Name of the file}
}
\value{
AEPTable
}
\description{
Import AEP values from user table in Excel
}
