% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeParseUHS.R
\name{safeParseUHS}
\alias{safeParseUHS}
\title{Safely parse UHS column names for (POE, Tn)}
\usage{
safeParseUHS(col_labels)
}
\arguments{
\item{col_labels}{Character vector of measure column names from meltdown}
}
\value{
A list of two numeric vectors: \code{(POE, Tn)}
}
\description{
Given a character vector of column names (e.g. \code{"0.09516~PGA"}),
returns two numeric vectors \code{(POE, Tn)}. If something doesn't parse,
those entries are \code{NA}.
}
\keyword{internal}
