% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShearModelParameters.R
\name{ShearModelParameters}
\alias{ShearModelParameters}
\title{ShearModelParameters}
\format{
A data frame with 20 rows and 10 variables:
\describe{
\item{ModelID}{Character: Model ID}
\item{GroupID}{Character: Group ID}
\item{NameID}{Character: NameID}
\item{AuthorID}{Character: AuthorID}
\item{emin}{Double: Min Void Ratio}
\item{emax}{Double: Max Void Ratio}
\item{A}{Double: A}
\item{Ce}{Double: Ce}
\item{n}{Double: Ce}
\item{UN}{Character: Go Units}
}
}
\usage{
data(ShearModelParameters)
}
\description{
This dataset contains hear Model Parameters from Ishihara (1996)
}
