% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDtoVs30.R
\name{SIDtoVs30}
\alias{SIDtoVs30}
\title{Convert Site Class to Vs30 in m/s}
\usage{
SIDtoVs30(SID = NULL)
}
\arguments{
\item{SID}{Site Class}
}
\value{
Vs30 in m/s
}
\description{
Convert Site Class to Vs30 in m/s
}
\examples{
SIDtoVs30("A")

SIDtoVs30("AB")

SIDtoVs30(c("BC","C","CD","D"))


}
