% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteClass.R
\name{SiteClass}
\alias{SiteClass}
\title{Site Class Data}
\format{
A data frame with 9 rows and 4 variables:
\describe{
\item{SC}{Character: Site Class}
\item{Description}{Character: Category description}
\item{Vs30 (m/s)}{Character: Vs30 in m/s}
\item{Vs30 (ft/s)}{Character: Vs30 in ft/s}
}
}
\usage{
data(SiteClass)
}
\description{
This dataset contains Site Class data from ASCE 7-22
}
