% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteTable.R
\name{SiteTable}
\alias{SiteTable}
\title{SiteTable Data}
\format{
A data frame with 255614 rows and 38 variables:
\describe{
\item{Hs}{Double: Height in m.}
\item{Hw}{Double: Water Table in m.}
\item{NL}{Integer: Number of Layers}
\item{Z500}{Double: Depth to 500 m/s in m.}
\item{Z1000}{Double: Depth to 1000 m/s in m.}
\item{SID}{String: Site ID}
\item{Go}{Double: Shear Modulus in kPa.}
\item{mo}{Double: Shear Modulus Exponent.}
\item{Ts}{Double: Fundamental Period in s.}
\item{VSo}{Double: Shear Wave Velocity in m/s.}
\item{VS30}{Double: Shear Wave Velocity at 30 m in m/s.}
\item{UID}{String: USCS ID}
\item{Gravels}{Double: Percentage of Gravels.}
\item{Sands}{Double: Percentage of Sands.}
\item{Fines}{Double: Percentage of Fines.}
\item{Clays}{Double: Percentage of Clays.}
\item{Silts}{Double: Percentage of Silts.}
\item{Organic}{Double: Percentage of Organic.}
\item{Water}{Double: Percentage of Water.}
\item{GP}{Double: GP Fraction.}
\item{GM}{Double: GM Fraction.}
\item{GC}{Double: GC Fraction.}
\item{GW}{Double: GW Fraction.}
\item{SP}{Double: SP Fraction.}
\item{SM}{Double: SM Fraction.}
\item{SC}{Double: SC Fraction.}
\item{SW}{Double: SW Fraction.}
\item{CL}{Double: CL Fraction.}
\item{ML}{Double: ML Fraction.}
\item{OL}{Double: OL Fraction.}
\item{PT}{Double: PT Fraction.}
\item{OH}{Double: OH Fraction.}
\item{MH}{Double: MH Fraction.}
\item{CH}{Double: CH Fraction.}
\item{POP}{Double: Overpressure in Kpa.}
\item{POP_Units}{String: Overpressure Units.}
\item{Go_Units}{String: Shear Modulus Units.}
\item{Vs_Units}{String: Shear Wave Velocity Units.}
}
}
\usage{
data(SiteTable)
}
\description{
This dataset contains simulations of different site conditions
}
