% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyColsWithDefaults.R
\name{unifyColsWithDefaults}
\alias{unifyColsWithDefaults}
\title{Force a data.table to have a fixed set of columns, filling defaults}
\usage{
unifyColsWithDefaults(dt, colNames, defaults = list())
}
\arguments{
\item{dt}{A data.table to unify}

\item{colNames}{Character vector of the final columns you want}

\item{defaults}{A named list of default values for specific columns, e.g.
\code{list(AF=1, sdLnAF=NA_real_)}. If a column in \code{colNames} is missing, we
fill it with \code{defaults[[col]]} if present, or NA if not.}
}
\value{
The same dt, now guaranteed to have columns in \code{colNames} with
defaults as specified.
}
\description{
For each desired column in \code{colNames}, if it's missing in \code{dt}, creates it
with a default value if specified in \code{defaults}, or NA otherwise. Also drops
any extra columns not in \code{colNames}. Reorders columns to match \code{colNames}.
}
