% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vs30toSID.R
\name{Vs30toSID}
\alias{Vs30toSID}
\title{Convert Vs30 in m/s to Site Class}
\usage{
Vs30toSID(Vs30)
}
\arguments{
\item{Vs30}{Double/Vector. Vs30 in m/s}
}
\value{
Character vector with Site Class.
}
\description{
Convert Vs30 in m/s to Site Class
}
\examples{

Vs30toSID(1500)

Vs30toSID(c(120,790,3000,455))

}
