% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Title}
\usage{
buildGMDP(
  path,
  IDo = "00000000",
  engine = "openquake",
  vs30 = NULL,
  vref,
  quantile_AF = "mean",
  POE_TARGET = NULL
)
}
\arguments{
\item{path}{character}

\item{IDo}{character Output IDo}

\item{engine}{character c("openquake","user")}

\item{vs30}{numeric Vs30 Step}

\item{vref}{Vs30 in m/s}

\item{quantile_AF}{character c("mean",0.16,0.50,0.84)}

\item{POE_TARGET}{numeric Target Probability of Exceedance}
}
\value{
list
}
\description{
Title
}
