% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Title}
\usage{
buildGMDP(
  path,
  IDo = "00000000",
  ITo,
  Vref = 760,
  engine = "openquake",
  Vs30_STEP = 25
)
}
\arguments{
\item{path}{Path to GMDP}

\item{IDo}{Output IDo}

\item{ITo}{Invesigation Time. ITo=50 yr}

\item{Vref}{Vs30 in m/s}

\item{engine}{c("openquake","user")}

\item{Vs30_STEP}{Vs30 Step}
}
\value{
list
}
\description{
Title
}
