% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Title}
\usage{
buildGMDP(path, IDo = "00000000", engine = "openquake", vs30 = NULL, vref)
}
\arguments{
\item{path}{character}

\item{IDo}{character Output IDo}

\item{engine}{character c("openquake","user")}

\item{vs30}{numeric Vs30 Step}

\item{vref}{Vs30 in m/s}
}
\value{
list
}
\description{
Title
}
