% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Title}
\usage{
buildGMDP(
  path,
  ID = "00000000",
  TRo = c(100, 200, 475, 500, 1000, 2000, 2475, 2500, 5000, 10000),
  ITo = 50,
  Vs30o = 760,
  Vs30_STEP = 25,
  configFile = "gmdp.ini"
)
}
\arguments{
\item{path}{Path to GMDP}

\item{ID}{Output ID}

\item{TRo}{Return Periods}

\item{ITo}{Invesigation Time. ITo=50 yr}

\item{Vs30o}{Vs30 in m/s}

\item{Vs30_STEP}{Vs30 Step}

\item{configFile}{GMDP config file}
}
\value{
GMDP
}
\description{
Title
}
