% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMDP.R
\name{buildGMDP}
\alias{buildGMDP}
\title{Title}
\usage{
buildGMDP(path, IDo = "00000000", engine = "openquake", Vs30 = NULL, Vref)
}
\arguments{
\item{path}{character}

\item{IDo}{character Output IDo}

\item{engine}{character c("openquake","user")}

\item{Vs30}{numeric Vs30 Step}

\item{Vref}{Vs30 in m/s}

\item{siteResponse}{boolean}
}
\value{
list
}
\description{
Title
}
