% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.Kh.R
\name{buildTable.Kh}
\alias{buildTable.Kh}
\title{Build Seismic Coefficient Table (Kmax) in g}
\usage{
buildTable.Kh(
  x,
  Tso,
  Dao,
  size = 12,
  po = c(0.16, 0.5, 0.84),
  engine = "flextable",
  TRo = c(500, 1000, 2500, 5000, 10000),
  Vs30o = NULL,
  SIDo = NULL
)
}
\arguments{
\item{x}{GMDP object}

\item{Tso}{Ts in seconds}

\item{Dao}{Da in cm}

\item{size}{font size}

\item{po}{quantiles}

\item{engine}{c("flextable")}

\item{TRo}{return periods in years}

\item{Vs30o}{Vs30 in m/s}

\item{SIDo}{Site ID}
}
\value{
Table
}
\description{
Build Seismic Coefficient Table (Kmax) in g
}
