% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.Sa.R
\name{buildTable.Sa}
\alias{buildTable.Sa}
\title{Title}
\usage{
buildTable.Sa(
  x,
  Tno,
  TRo,
  size = 12,
  po = NULL,
  engine = "flextable",
  Vs30o = NULL,
  tagUnits = TRUE
)
}
\arguments{
\item{x}{GMDP object}

\item{Tno}{Period in seconds}

\item{TRo}{Return Period in years}

\item{size}{Font size}

\item{po}{c("mean",0.16,0.50,0.84)}

\item{engine}{c("flextable","kableExtra")}

\item{Vs30o}{Vs30 in m/s}

\item{tagUnits}{boolean}
}
\value{
Table
}
\description{
Title
}
