% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.AF.TR.R
\name{fitModel.AF.TR}
\alias{fitModel.AF.TR}
\title{Amplification Factor AF, from Stewart Et Al. 2016}
\usage{
fitModel.AF.TR(.x, pga, Tn, vs30, vref, q = 0.5, Vl = 200, Vu = 2000)
}
\arguments{
\item{.x}{data.table. Input data.table}

\item{pga}{Double. PGA in g}

\item{Tn}{Double. Natural Period}

\item{vs30}{Double. Vs30 in m/s}

\item{vref}{Double. Reference Vs30 in m/s}

\item{q}{Double. quantile}

\item{Vl}{Double. Lower limit Vs30 in m/s}

\item{Vu}{Double. Upper limit Vs30 in m/s}
}
\value{
data.table
}
\description{
Amplification Factor AF, from Stewart Et Al. 2016
}
