% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.AF.TR.R
\name{fitModel.AF.TR}
\alias{fitModel.AF.TR}
\title{Amplification Factor AF, from Stewart Et Al. 2016}
\usage{
fitModel.AF.TR(.x, p, Tn, Vs30, Vref, Vl = 200, Vu = 2000)
}
\arguments{
\item{.x}{GMDP object}

\item{p}{Double. Period in seconds}

\item{Tn}{Double. Natural Period}

\item{Vs30}{Double. Vs30 in m/s}

\item{Vref}{Double. Reference Vs30 in m/s}

\item{Vl}{Double. Lower limit Vs30 in m/s}

\item{Vu}{Double. Upper limit Vs30 in m/s}
}
\value{
data.table
}
\description{
Amplification Factor AF, from Stewart Et Al. 2016
}
