% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Kmax.Ts.R
\name{fitModel.Kmax.Ts}
\alias{fitModel.Kmax.Ts}
\title{Title}
\usage{
fitModel.Kmax.Ts(.data, Tso, Dao, model = "rf", OSF = 0.3)
}
\arguments{
\item{.data}{data.table}

\item{Tso}{double}

\item{Dao}{double}

\item{model}{string c("nlm","rf")}

\item{OSF}{double}
}
\description{
Title
}
