% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Sa.TR.R
\name{fitModel.Sa.TR}
\alias{fitModel.Sa.TR}
\title{Linear regression model for Sa in terms of TR}
\usage{
fitModel.Sa.TR(x, TRmin = 100, TRmax = 10000)
}
\arguments{
\item{x}{data.table with columns Sa and TR}

\item{TRmin}{minimum TR value to fit}

\item{TRmax}{maximum TR value to fit}
}
\value{
data table with columns a, b, c, sdLnA, R2, MSE, RMSE, fit
}
\description{
Linear regression model for Sa in terms of TR
}
