% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.Td.R
\name{fitModel.Td}
\alias{fitModel.Td}
\title{Title}
\usage{
fitModel.Td(
  Hmax,
  lo = NULL,
  Bmax = NULL,
  Bo = NULL,
  s = NULL,
  .newdata,
  level = 0.5,
  regression = "qrf"
)
}
\arguments{
\item{Hmax}{Double. Height in m}

\item{lo}{Double. Truncation ratio}

\item{Bmax}{Double. Maximum width in m}

\item{Bo}{Double. TSF Crest in m}

\item{s}{Double. Slope S=1/tan(beta)}

\item{.newdata}{Data frame. New data to predict}

\item{.level}{Double. Confidence level (0-1)}

\item{.regression}{String. Regression method}
}
\value{
List
}
\description{
Title
}
