% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.oqAEP.R
\name{importModel.oqAEP}
\alias{importModel.oqAEP}
\title{Import Annual Exceedance Probabiliies (AEP) from openquake zip files.}
\usage{
importModel.oqAEP(path, ITo)
}
\arguments{
\item{path}{character. Path to the folder containing the hazard and quantile curves}

\item{ITo}{numeric}
}
\value{
A data.table with the following columns:
}
\description{
Import Annual Exceedance Probabiliies (AEP) from openquake zip files.
}
