% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.oqAEP.R
\name{importModel.oqAEP}
\alias{importModel.oqAEP}
\title{Import Annual Exceedance Probabiliies (AEP) from openquake zip files.}
\usage{
importModel.oqAEP(path, vref)
}
\arguments{
\item{path}{character. Path to the folder containing the hazard and quantile curves}

\item{vref}{numeric. Reference Vs30 in m/s}
}
\value{
A data.table with the following columns:
}
\description{
Import Annual Exceedance Probabiliies (AEP) from openquake zip files.
}
