% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.oqRMw.R
\name{importModel.oqRMw}
\alias{importModel.oqRMw}
\title{Import disaggregation hazard (R-MW-p) from openquake}
\usage{
importModel.oqRMw(path, gmdp.ini = NULL)
}
\arguments{
\item{path}{Path to the folder containing the hazard and quantile curves}

\item{gmdp.ini}{List with the following elements:}
}
\value{
A data.table with the following columns:
}
\description{
Import disaggregation hazard (R-MW-p) from openquake
}
