% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.oqRMw.R
\name{importModel.oqRMw}
\alias{importModel.oqRMw}
\title{Import Disaggregation (R-Mw-p) from OpenQuake}
\usage{
importModel.oqRMw(path, ITo, vref)
}
\arguments{
\item{path}{Character. Folder with \code{.zip} or direct \verb{Mag_Dist*} CSV}

\item{ITo}{Numeric. Investigation time}

\item{vref}{Numeric. Vs30 for reference (not mandatory)}
}
\value{
data.table or NULL if no valid data found. Columns:
\itemize{
\item Mw, R, POE, p, Tn, AEP, TR, IT
}
}
\description{
Reads files named "Mag_Dist" (not "TRT") from the .zip or folder. If a
"Mag_Dist-mean" is present, we pick that first. Skips the first line
(metadata), treats second line as column headers, renames (mag,dist,poe)->(Mw,R,POE),
and (rlz|mean)->p. Then calculates AEP=POE/ITo, TR=1/AEP, Tn from "imt" if present.
}
