% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModel.oqUHS.R
\name{importModel.oqUHS}
\alias{importModel.oqUHS}
\title{Import UHS Curves from OpenQuake Outputs}
\usage{
importModel.oqUHS(path)
}
\arguments{
\item{path}{Character. Directory with UHS csv or zip}
}
\value{
A data.table with columns:
\itemize{
\item \strong{lon, lat, depth} (if depth is found)
\item \strong{rate, Tn, Sa, p, ITo}
}
}
\description{
Reads "uhs*.csv" or "uhs-csv.zip" from \code{path}, skipping the first line
(metadata) and treating the second line as the column header. Then each
column name like "0.095160~PGA" is melted, with "exceed_tag" parsed by
\code{.parseUHScolumn()} => \verb{(rate, Tn)}.
}
