% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTable.R
\name{renderTable}
\alias{renderTable}
\title{Title}
\usage{
renderTable(x, fontname = "Helvetica", size = 10)
}
\arguments{
\item{x}{Data frame.}

\item{fontname}{String. Font name.}

\item{size}{Numeric. Font size.}
}
\value{
flextable
}
\description{
Title
}
