% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTable.R
\name{renderTable}
\alias{renderTable}
\title{Title}
\usage{
renderTable(
  .x,
  fontname = "Helvetica",
  fontsize.header = "11px",
  fontsize.body = "10px",
  border.color = "blue"
)
}
\arguments{
\item{.x}{Data frame.}

\item{fontname}{String. Font name.}

\item{fontsize.header}{String Font size in px.}

\item{fontsize.body}{String Font size in px.}

\item{border.color}{String. Border color.}
}
\value{
gt object
}
\description{
Title
}
